Copy the EXEPACK block farther away when compression is slight.
In the case of light compression, the end of the compressed data
(dest_len) may lie in the middle of the decompression stub. Copying the
stub to dest_len would clobber the stub with itself while it is still
running. So in this case, we move the block even farther, to just past
the end of itself.

This allows for "compressing" executables that get larger after the
compression operation, like when you run exepack on the output of
exepack.