Make write_exe adjust its header values.
Otherwise, reading hello.exe and writing it again would have e.g.
e_cparhdr pointing into the middle of the header padding, because we
were doing our own padding to a multiple of 512 bytes without adjusting
the header values to match. The problem was hidden because pack and
unpack produce EXEs with headers that were compatible with what
write_exe was doing before.