Use a stack size of 128 bytes in compressed executables.
I believe our stub itself only uses at most 4 bytes of stack, but
interrupts can also make use of the stack. 128 bytes is the stack size
used by the Microsoft compressors, so that should be a reasonable size
to use.

The compressor also now tries to leave e_ss aligned with the top of the
allocated stack area; i.e., e_sp = 128 in most cases. This is in
contrast to the previous behavior of always shifting as many bits into
e_sp as possible. I made this change because 1) that's what the
Microsoft compressors do, and 2) it may be a little safer in case of a
stack overflow, which will wrap and start writing to the top of the
segment rather than overwriting program code. Bits are still shifted
from e_ss to e_sp in extreme cases where it is necessary.