Handle e_minalloc correctly in pack and unpack.
Formerly, we copied e_minalloc unchanged from input to output in both
directions. After investigation of how UNP and MASM 4.00 handle this
field, I understand it better. The key idea is that e_minalloc is like
the BSS segment in a Unix executable. It's an uninitialized block of
memory that is effectively part of the program image, but only its size
is stored in the EXE file.

Generally we want to keep the quantity
  round_up(size) + e_minalloc*16
constant, where round_up rounds to a multiple of 16 bytes, and size is
the file size from the e_cblp and e_cp fields, minus e_cparhdr*16.

In the compression direction, we may have to increase the total size
beyond the sum above (by adjusting e_minalloc upward) to account for the
overhead of the decompression stub and its preliminary copy operation.

In the decompression direction, we can maintain the invariant by
adjusting e_minalloc, except in case of underflow or overflow.
(Underflow should not happen with properly compressed files.)