Avoid overflow in ceil(exepack_size/16) in stub.
I noticed this while trying to run
tests/maxlen_maxrelocs_compressible.packed.exe, which has an
exepack_size of 0xffff. The program doesn't run in DOSBox anyway, unless
I reduce the size of the uncompressed data in the test. (I'm guessing it
has something to do with a too-large stack pointer.) But if I altered
the test to use `len = 0x1f000`, then DOSBox would run it, and before
this change crash due to the overflow.