Mark the places where the stub may overflow.
We cannot statically guarantee these additions will not overflow, mainly
because they are relative to mem_start, and we don't know in advance
where where the OS might have located us.

It looks like the loaders of FreeDOS and DOSBox similarly ignore
overflow:
https://github.com/FDOS/kernel/blob/ke2043/kernel/task.c#L777
https://sourceforge.net/p/dosbox/code-0/HEAD/tree/dosbox/tags/RELEASE_0_74_3/src/dos/dos_execute.cpp#l401

It would be nice to exit with an error message if overflow is detected.
But the error message "Packed file is corrupt" would not be appropriate
(because unpacking may in fact succeed if the program is loaded at a
lower address), and it would be hard to fit in another error message
under the byte limit.